Changes how fields are handled in external search
The searchable fields and sortable fields have been moved from
the indexer to the base config and is not provided to the indexer
as a parameter. The main purpose of this change it to allow for the
service itself to know about the indexed fields.

It might be beneficial to adopte the search input to this new format
to allow better handling of language-dependent fields.