Adds a resync feature to the ext. bindings
The paths used in Infoglue for external bindings are static text since
there is no live connection to the external data source. This new feature
allows the external binding view to resubmit names for the events in the
current list.

When the view is loaded all current IDs are passed to the external binding
view and the view may call the resyncNames() function in the parent view
that will update the names in the current list (to the right).

This will also resolve any out-of-sync messages in the parent view.