Fixes a problem where the main frame was renamed
In the content view a variabel called 'name' was defined in the global
scope. In modern browsers this caused the frame to be renamed according to
the value of the variabel. Which in turn caused other operations to be
unable to locate the main frame. The variable is now renamed.